+++ packages/model/stop_daemons.go
@@ -3,7 +3,7 @@ package model
 import "time"
 
 type StopDaemon struct {
-	StopTime int32 `gorm:"not null"`
+	StopTime int64 `gorm:"not null"`
 }
 
 func (sd *StopDaemon) TableName() string {
@@ -22,6 +22,6 @@ func (sd *StopDaemon) Get() error {
 	return DBConn.First(sd).Error
 }
 func SetStopNow() error {
-	stopTime := &StopDaemon{StopTime: int32(time.Now().Unix())}
+	stopTime := &StopDaemon{StopTime: time.Now().Unix()}
 	return stopTime.Create()
 }
